Added better Save support
The app now can save tasks even through being completely closed. This is
done by saving the tasks to a private file. There are currently a lot of
Log.d statements (print statements) that I used to find errors. So now
we can make and save tasks. We need to add a way to check  them now
which will come from editing the listview stuff. I'll look in to that